+++ src/BitmapFont.ts
@@ -39,7 +39,7 @@ export interface BitmapFontParameterObject {
 /**
  * ラスタ画像によるフォント。
  */
-export class BitmapFont implements Font {
+export class BitmapFont extends Font {
 	surface: Surface;
 	defaultGlyphWidth: number;
 	defaultGlyphHeight: number;
@@ -52,6 +52,7 @@ export class BitmapFont implements Font {
 	 * @param param `BitmapFont` に設定するパラメータ
 	 */
 	constructor(param: BitmapFontParameterObject) {
+		super();
 		this.surface = SurfaceUtil.asSurface(param.src);
 		this.map = param.map;
 		this.defaultGlyphWidth = param.defaultGlyphWidth;