+++ src/Camera.ts
@@ -184,7 +184,7 @@ export class Camera2D extends Object2D implements Camera {
 	 * @private
 	 */
 	_applyTransformToRenderer(renderer: Renderer): void {
-		if (this.angle || this.scaleX !== 1 || this.scaleY !== 1) {
+		if (this.angle || this.scaleX !== 1 || this.scaleY !== 1 || this.anchorX != null || this.anchorY != null) {
 			// Note: this.scaleX/scaleYが0の場合描画した結果何も表示されない事になるが、特殊扱いはしない
 			renderer.transform(this.getMatrix()._matrix);
 		} else {
@@ -199,8 +199,28 @@ export class Camera2D extends Object2D implements Camera {
 	 */
 	_updateMatrix(): void {
 		// カメラの angle, x, y はエンティティと逆方向に作用することに注意。
-		if (this.angle || this.scaleX !== 1 || this.scaleY !== 1) {
-			this._matrix.updateByInverse(this.width, this.height, this.scaleX, this.scaleY, this.angle, this.x, this.y);
+		if (this.anchorX != null && this.anchorY != null) {
+			this._matrix.updateByInverseWithAnchor(
+				this.width,
+				this.height,
+				this.scaleX,
+				this.scaleY,
+				this.angle,
+				this.x,
+				this.y,
+				this.anchorX,
+				this.anchorY
+			);
+		} else if (this.angle || this.scaleX !== 1 || this.scaleY !== 1) {
+			this._matrix.updateByInverse(
+				this.width,
+				this.height,
+				this.scaleX,
+				this.scaleY,
+				this.angle,
+				this.x,
+				this.y
+			);
 		} else {
 			this._matrix.reset(-this.x, -this.y);
 		}