+++ src/Font.ts
@@ -1,10 +1,12 @@
 import { Destroyable } from "./Destroyable";
 import { Glyph } from "./Glyph";
+import { TextMetrix } from "./TextMetrix";
+import { Util } from "./Util";
 
 /**
  * フォント。
  */
-export interface Font extends Destroyable {
+export abstract class Font implements Destroyable {
 	/**
 	 * フォントサイズ。
 	 *
@@ -19,5 +21,48 @@ export interface Font extends Destroyable {
 	 *
 	 * @param code 文字コード
 	 */
-	glyphForCharacter(code: number): Glyph;
+	abstract glyphForCharacter(code: number): Glyph;
+
+	abstract destroy(): void;
+
+	abstract destroyed(): boolean;
+
+	/**
+	 * 対象の文字列を一行で描画した際の計測情報を返す。
+	 *
+	 * @param text 文字列
+	 */
+	measureText(text: string): TextMetrix {
+		let width = 0;
+		let actualBoundingBoxLeft = 0;
+		let actualBoundingBoxRight = 0;
+		let lastGlyph: Glyph;
+
+		for (let i = 0; i < text.length; i++) {
+			const code = Util.charCodeAt(text, i);
+			if (!code)
+				continue;
+
+			const glyph = this.glyphForCharacter(code);
+			if (!glyph || glyph.x < 0 || glyph.y < 0 || glyph.width < 0 || glyph.height < 0)
+				continue;
+
+			if (i === 0) {
+				actualBoundingBoxLeft = -glyph.offsetX;
+			}
+
+			lastGlyph = glyph;
+			width += glyph.advanceWidth;
+		}
+
+		if (lastGlyph) {
+			actualBoundingBoxRight = width + lastGlyph.offsetX + lastGlyph.width - lastGlyph.advanceWidth;
+		}
+
+		return {
+			width,
+			actualBoundingBoxLeft,
+			actualBoundingBoxRight
+		};
+	}
 }